# docs/development/figures/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Cloudogu User Management",
    "description": "Api description for cloudogu User Management",
    "version": "0.1"
  },
  "servers": [
    {
      "url": "https://192.168.56.2/usermgt"
    }
  ],
  "paths": {
...
    "/api/users": {
      "get": {
        "description": "Get a list with all users.",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/inline_response_200_1"
                },
                "examples": {
                  "0": {
                    "value": "{\"entries\":[{\"displayName\":\"admin\",\"givenname\":\"admin\",\"mail\":\"admin@ces.local\",\"memberOf\":[\"admin2\",\"cesManager\"],\"password\":\"__dummypassword\",\"pwdReset\":false,\"surname\":\"admin\",\"username\":\"admin\"},{\"displayName\":\"admin2\",\"givenname\":\"admin2\",\"mail\":\"admin2@test.de\",\"memberOf\":[],\"password\":\"__dummypassword\",\"pwdReset\":true,\"surname\":\"admin2\",\"username\":\"admin2\"},{\"displayName\":\"test\",\"givenname\":\"test\",\"mail\":\"test@cloudogu.com\",\"memberOf\":[],\"password\":\"__dummypassword\",\"pwdReset\":false,\"surname\":\"test\",\"username\":\"testuser\"}],\"start\":0,\"limit\":20,\"totalEntries\":3}"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "description": "Create a new user",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/api_users_body"
              },
              "examples": {
                "0": {
                  "value": "{\n    \"memberOf\": [],\n    \"username\": \"mynewusertocreate2\",\n    \"givenname\": \"mynewusertocreate\",\n    \"surname\": \"mynewusertocreate\",\n    \"displayName\": \"mynewusertocreate\",\n    \"mail\": \"mynewusertocreate2@test.de\",\n    \"password\": \"mynewusertocreate\",\n    \"pwdReset\": true\n}"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Auto generated using Swagger Inspector"
          }
        }
      }
    },
    "/api/users/import": {
      "post": {
        "summary": "Import Users from CSV",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "CSV file to be imported"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful import",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportError"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/users/import/{importID}": {
      "get": {
        "summary": "Get Import Result",
        "parameters": [
          {
            "name": "importID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the import result."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            },
            "headers": {
              "Content-Disposition": {
                "description": "File attachment header",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "Invalid UUID",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Result not found",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/users/import/summaries": {
      "get": {
        "summary": "Get summaries",
        "parameters": [
          {
            "name": "start",
            "in": "query",
            "description": "Start index for pagination (default: 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return per page (default: 10, maximum: 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResultListSummary"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/users/{username}/groups/{groupname}": {
      "post": {
        "description": "Adds a group member.",
        "parameters": [
          {
            "name": "groupname",
            "in": "path",
            "description": "Name of the group",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "username",
            "in": "path",
            "description": "Name of the user",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          },
          "409": {
            "description": "Error -> User is already member of the group"
          }
        }
      },
      "delete": {
        "description": "Removes a member from a group.",
        "parameters": [
          {
            "name": "groupname",
            "in": "path",
            "description": "Name of the group",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "username",
            "in": "path",
            "description": "Name of the user",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "inline_response_200_1_entries": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "pwdReset": {
            "type": "boolean"
          },
          "mail": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "givenname": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "memberOf": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "username": {
            "type": "string"
          }
        },
        "example": null
      },
      "api_users_body": {
        "type": "object",
        "properties": {
          "password": {
            "type": "string"
          },
          "pwdReset": {
            "type": "boolean"
          },
          "mail": {
            "type": "string"
          },
          "givenname": {
            "type": "string"
          },
          "surname": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "memberOf": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "inline_response_200_1": {
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/inline_response_200_1_entries"
            }
          },
          "start": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "totalEntries": {
            "type": "integer"
          }
        }
      },
...
      "Result": {
        "type": "object",
        "properties": {
          "importID": {
            "type": "string",
            "description": "The ID of the import result."
          },
          "filename": {
            "type": "string",
            "description": "The filename of the import result."
          },
          "timestamp": {
            "type": "integer",
            "format": "int64",
            "description": "The timestamp of the import result."
          },
          "created": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/api_users_body"
            },
            "description": "List of created users."
          },
          "updated": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/api_users_body"
            },
            "description": "List of updated users."
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportError"
            },
            "description": "List of import errors."
          }
        }
      },
      "ImportError": {
        "type": "object",
        "properties": {
          "errorCode": {
            "type": "integer",
            "description": "The error code representing the type of error."
          },
          "lineNumber": {
            "type": "integer",
            "description": "The line number of the import error."
          },
          "message": {
            "type": "string",
            "description": "The error message describing the reason for the error."
          },
          "params": {
            "type": "object",
            "properties": {
              "columns": {
                "type": "array",
                "items": {
                  "type": "string"
                },
                "description": "List of affected columns in case of an error."
              }
            }
          }
        },
        "required": ["errorCode", "lineNumber", "message"]
      },
      "PagedResultListSummary": {
        "type": "object",
        "properties": {
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ImportSummary"
            }
          },
          "start": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "totalEntries": {
            "type": "integer"
          }
        }
      },
      "ImportSummary": {
        "type": "object",
        "properties": {
          "importID": {
            "type": "string",
            "description": "The ID of the import."
          },
          "filename": {
            "type": "string",
            "description": "The filename of the import."
          },
          "created": {
            "type": "integer",
            "description": "The number of users created during the import."
          },
          "updated": {
            "type": "integer",
            "description": "The number of users updated during the import."
          },
          "skipped": {
            "type": "integer",
            "description": "The number of rows skipped during the import."
          }
        }
      }
    },
    "securitySchemes": {
      "basic": {
        "type": "http",
        "scheme": "basic"
      }
    }
  }
}